[chore] relocate login files
