+++ src/app/(web)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import "@/styles/globals.css";
 import { ReactQueryProvider } from "./_providers";
-import AuthProvider from "@/features/auth/authProvider";
+import AuthProvider from "@/features/auth/providers/authProvider";
 import { ModalContainer } from "@/features/modal/components";
 import { ToastContainer } from "@/features/toast/components";
 import { Header } from "./_components";
