--- src/features/auth/login/index.ts
@@ -1,4 +1,0 @@
-export { default as LoginModal } from "./loginModal";
-export { default as LoginForm } from "./loginForm";
-export { default as useLogin } from "./useLogin";
-export { loginSchema, type LoginFormDataType } from "./loginSchema";
